+++ src/App.jsx
@@ -1,12 +1,10 @@
-import About from "./components/About";
+import About from "./components/About/About";
 import Contact from "./components/Contact";
-import Hero from "./components/Hero/Hero";
 import Projects from "./components/Projects";
 
 const App = () => {
   return (
     <div>
-      <Hero />
       <About />
       <Projects />
       <Contact />
