+++ src/components/About/About.jsx
@@ -1,4 +1,4 @@
-import "./Hero.css";
+import "./About.css";
 
 const Hero = () => {
   return (
